update the project
I do the prediction of my model and see how is the performance of my model. Even though the Loss in training looks good we still need prediction to verify it. I did the prediction with the test dataset and it looks fine to me. I also show the result of the prediction and make a graph of the head 1000 result of my prediction.